dev(full-code-rewrite-and-improvement): Docstrings, song/util.py and user/util.py
